Simplify and correct perl module generation. texi2any.pl: better defaults for variables set by configure.

+++ tp/maintain/regenerate_perl_module_files.sh
@@ -4,6 +4,8 @@
 
 # This regenerate files in t/include_dir/
 make
+PACKAGE=`grep '^PACKAGE = ' Makefile | sed 's/^PACKAGE = //'`
+[ z"$PACKAGE" = 'z' ] && exit 1
 rm -rf LocaleData
 
 for dir in po po_document; do
@@ -11,7 +13,8 @@ for dir in po po_document; do
   mkdir $dir
   cp maintain/Makefile_perl_po $dir/Makefile
   grep '^tp\/' ../$dir/POTFILES.in | sed -e 's/^tp/../' > $dir/POTFILES.in
-  cp maintain/package_name $dir/PACKAGE
+  echo "# File automatically generated from texinfo files"
+  echo "PACKAGE = $PACKAGE" > $dir/PACKAGE
   cat ../$dir/Makevars >> $dir/PACKAGE
   linguas_str=
   for lingua in `cat ../$dir/LINGUAS`; do
